Refactor authentication client
This is a small refactoring that handles validation earlier and holds on to username and password as properties of the class, which reduces the need to pass these values around. I began this refactoring when doing the work to build a tiny authN CLI yesterday, to increase my own understanding of what was going on behind the scenes. As part of the refactoring, I also made the tests smaller and more granular.